sources: Code simplification of simple_switch_13.py
simple_switch_13.py of Ryu source is often updated.
So simple_switch_13.py is difficult as a description source.
example_switch_13.py exists in Ryu.
example_switch_13.py is the file that simplifies simple_switch_13.py.
To reflect the contents of the example_switch_13.py to simple_switch_13.py of Ryu-Book.

Signed-off-by: Shinpei Muraoka <shinpei.muraoka@gmail.com>